+++ examples/js/keyboard-nav/enter-to-select-row-with-nav-table.js
@@ -1,6 +1,4 @@
 /* eslint max-len: 0 */
-/* eslint no-alert: 0 */
-/* eslint no-console: 0 */
 import React from 'react';
 import { BootstrapTable, TableHeaderColumn } from 'react-bootstrap-table';
 
@@ -19,23 +17,18 @@ function addProducts(quantity) {
   }
 }
 
-addProducts(50);
+addProducts(5);
 
-const options = {
-  onRowClick: function(row, columnIndex, rowIndex, e) {
-    alert(`You click row id: ${row.id}, column index: ${columnIndex}, row index: ${rowIndex}`);
-    console.log(e);
-  },
-  onRowDoubleClick: function(row, e) {
-    alert(`You double click row id: ${row.id}`);
-    console.log(e);
-  }
-};
-
-export default class SingleSelectTable extends React.Component {
+export default class SimpleNavTable extends React.Component {
   render() {
+    const selectRow = {
+      mode: 'checkbox'
+    };
+    const keyBoardNav = {
+      enterToSelect: true
+    };
     return (
-      <BootstrapTable data={ products } options={ options } pagination>
+      <BootstrapTable data={ products } selectRow={ selectRow } keyBoardNav={ keyBoardNav }>
           <TableHeaderColumn dataField='id' isKey={ true }>Product ID</TableHeaderColumn>
           <TableHeaderColumn dataField='name'>Product Name</TableHeaderColumn>
           <TableHeaderColumn dataField='price'>Product Price</TableHeaderColumn>